requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
alembic==1.16.5
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.10.1
jinja2==3.1.6
python-multipart==0.0.20
apscheduler==3.11.0
httpx==0.28.1
fpdf2==2.8.4
pytest==8.4.2
pytest-asyncio==0.24.0
